fixes requested changes, moves all handlers into src/handlers/ folder
