+++ backend/serverless.yml
@@ -78,63 +78,63 @@ plugins:
 
 functions:
   hello:
-    handler: src/handler.hello
+    handler: src/handlers/handler.hello
     events:
       - http:
           method: get
           path: hello
   dbInit:
-    handler: src/db.init
+    handler: src/handlers/db.init
   dbDrop:
-    handler: src/db.drop
+    handler: src/handlers/db.drop
 
   # Folder endpoints
   getFolder:
-    handler: src/folder.getFolder
+    handler: src/handlers/folder.getFolder
     events:
       - http:
           method: get
           path: /api/v1/folder
   newFolder:
-    handler: src/folder.newFolder
+    handler: src/handlers/folder.newFolder
     events:
       - http:
           method: post
           path: /api/v1/folder
   updateFolder:
-    handler: src/folder.updateFolder
+    handler: src/handlers/folder.updateFolder
     events:
       - http:
           method: put
           path: /api/v1/folder
   deleteFolder:
-    handler: src/folder.deleteFolder
+    handler: src/handlers/folder.deleteFolder
     events:
       - http:
           method: delete
-          path: folder
+          path: /api/v1/folder
 
   # Collection endpoints
   getCollection:
-    handler: src/collection.get
+    handler: src/handlers/collection.get
     events:
       - http:
           method: get
           path: /api/v1/collection
   createCollection:
-    handler: src/collection.create
+    handler: src/handlers/collection.create
     events:
       - http:
           method: post
           path: /api/v1/collection
   updateCollection:
-    handler: src/collection.update
+    handler: src/handlers/collection.update
     events:
       - http:
           method: put
           path: /api/v1/collection
   deleteCollection:
-    handler: src/collection.remove
+    handler: src/handlers/collection.remove
     events:
       - http:
           method: delete
@@ -142,25 +142,25 @@ functions:
 
   # Question endpoints
   getQuestion:
-    handler: src/question.get
+    handler: src/handlers/question.get
     events:
       - http:
           method: get
           path: /api/v1/question
   createQuestion:
-    handler: src/question.create
+    handler: src/handlers/question.create
     events:
       - http:
           method: post
           path: /api/v1/question
   updateQuestion:
-    handler: src/question.update
+    handler: src/handlers/question.update
     events:
       - http:
           method: put
           path: /api/v1/question
   deleteQuestion:
-    handler: src/question.remove
+    handler: src/handlers/question.remove
     events:
       - http:
           method: delete
@@ -168,19 +168,19 @@ functions:
 
   # Question Option endpoints
   createQuestionOption:
-    handler: src/question_option.create
+    handler: src/handlers/question_option.create
     events:
       - http:
           method: post
           path: /api/v1/question_option
   updateQuestionOption:
-    handler: src/question_option.update
+    handler: src/handlers/question_option.update
     events:
       - http:
           method: put
           path: /api/v1/question_option
   deleteQuestionOption:
-    handler: src/question_option.remove
+    handler: src/handlers/question_option.remove
     events:
       - http:
           method: delete
@@ -188,13 +188,13 @@ functions:
 
   # Question User Response endpoints
   createQuestionUserResponse:
-    handler: src/question_user_response.create
+    handler: src/handlers/question_user_response.create
     events:
       - http:
           method: post
           path: /api/v1/question_user_response
   deleteQuestionUserResponse:
-    handler: src/question_user_response.remove
+    handler: src/handlers/question_user_response.remove
     events:
       - http:
           method: delete
@@ -202,42 +202,42 @@ functions:
 
   # websocket endpoints
   connect:
-    handler: src/ws/connect.handler
+    handler: src/handlers/websocket/connect.handler
     events:
       - websocket:
           route: $connect
   disconnect:
-    handler: src/ws/disconnect.handler
+    handler: src/handlers/websocket/disconnect.handler
     events:
       - websocket:
           route: $disconnect
   join:
-    handler: src/ws/join.handler
+    handler: src/handlers/websocket/join.handler
     events:
       - websocket:
           route: join
   createRoom:
-    handler: src/ws/createRoom.handler
+    handler: src/handlers/websocket/createRoom.handler
     events:
       - websocket:
           route: createRoom
   closeRoom:
-    handler: src/ws/closeRoom.handler
+    handler: src/handlers/websocket/closeRoom.handler
     events:
       - websocket:
           route: closeRoom
   submit:
-    handler: src/ws/submit.handler
+    handler: src/handlers/websocket/submit.handler
     events:
       - websocket:
           route: submit
   startQuestion:
-    handler: src/ws/startQuestion.handler
+    handler: src/handlers/websocket/startQuestion.handler
     events:
       - websocket:
           route: startQuestion
   endQuestion:
-    handler: src/ws/endQuestion.handler
+    handler: src/handlers/websocket/endQuestion.handler
     events:
       - websocket:
           route: endQuestion
@@ -268,7 +268,7 @@ functions:
 #            detail:
 #              state:
 #                - pending
-#      - cloudwatchLog: '/aws/lambda/hello'
+#      - cloudwatchLog: '/awebsocket/lambda/hello'
 #      - cognitoUserPool:
 #          pool: MyUserPool
 #          trigger: PreSignUp
