+++ backend/src/handlers/websocket/closeRoom.ts
